+++ Job/JobParameters/DefaultValuesProvider/FamilyCsvExport.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Pim\Bundle\EnhancedConnectorBundle\Job\JobParameters\DefaultValuesProvider;
+
+use Akeneo\Component\Batch\Job\JobInterface;
+use Akeneo\Component\Batch\Job\JobParameters\DefaultValuesProviderInterface;
+
+/**
+ * DefaultParameters for family CSV export
+ *
+ * @author    Philippe Mossière <philippe.mossiere@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+class FamilyCsvExport implements DefaultValuesProviderInterface
+{
+    /** @var DefaultValuesProviderInterface */
+    protected $simpleProvider;
+
+    /** @var array */
+    protected $supportedJobNames;
+
+    /**
+     * @param DefaultValuesProviderInterface $simpleProvider
+     * @param array                          $supportedJobNames
+     */
+    public function __construct(DefaultValuesProviderInterface $simpleProvider, array $supportedJobNames)
+    {
+        $this->simpleProvider    = $simpleProvider;
+        $this->supportedJobNames = $supportedJobNames;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDefaultValues()
+    {
+        $parameters = $this->simpleProvider->getDefaultValues();
+        $parameters['filePath'] = '/tmp/csv_family_pimgento_%datetime%.csv';
+        $parameters['labelLocale'] = null;
+
+        return $parameters;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function supports(JobInterface $job)
+    {
+        return in_array($job->getName(), $this->supportedJobNames);
+    }
+}