+++ PimEnhancedConnectorBundle.php
@@ -2,7 +2,7 @@
 
 namespace Pim\Bundle\EnhancedConnectorBundle;
 
-use Akeneo\Bundle\BatchBundle\Connector\Connector;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
  * Enhanced connector bundle.
@@ -11,6 +11,6 @@ use Akeneo\Bundle\BatchBundle\Connector\Connector;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class PimEnhancedConnectorBundle extends Connector
+class PimEnhancedConnectorBundle extends Bundle
 {
 }