+++ spec/Pim/Bundle/EnhancedConnectorBundle/Processor/Normalization/FamilyProcessorSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\EnhancedConnectorBundle\Processor;
+namespace spec\Pim\Bundle\EnhancedConnectorBundle\Processor\Normalization;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Repository\LocaleRepositoryInterface;
@@ -9,16 +9,11 @@ use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 /**
  * @author Romain Monceau <romain@akeneo.com>
  */
-class FamilyToFlatArrayProcessorSpec extends ObjectBehavior
+class FamilyProcessorSpec extends ObjectBehavior
 {
     function let(NormalizerInterface $transNormalizer, LocaleRepositoryInterface $localeRepository)
     {
         $this->beConstructedWith($transNormalizer, $localeRepository);
-    }
-
-    function it_is_a_configurable_step_element()
-    {
-        $this->beAnInstanceOf('Akeneo\Component\Batch\Item\AbstractConfigurableStepElement');
     }
 
     function it_is_a_processor()